Add .csv file containing updated mortality data
This .csv consists of a table containing the updated Puerto Rico mortality data for years 2015 (January) through 2018 (May), as provided by the P. R. Health Department. The table has a header and 6 distinct variables: Month, Day, [and values for years] 2015, 2016, 2017 and 2018. Each specific year-month-day combination has a corresponding cell containing the official count of deaths registered for that date, barring days for which such a count would be incomplete (0 for any date exceeding May 30th, 2017) or invalid (NA in the cells for February 29th in 2015, 2017 and 2018). Additional cells are included for each monthly count's total, average, maximum, minimum and median, with NA values for all months exceeding May, 2017.